requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ─── LLM & Embeddings ────────────────────────────────────────────────
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.1

# ─── Vector Store & Reranking ─────────────────────────────────────────
pinecone>=3.0.0
cohere>=5.0.0

# ─── Web Framework & UI ──────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
httpx>=0.25.0

# ─── Database & Cache ─────────────────────────────────────────────────
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0

# ─── Document Parsing ─────────────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ─── BM25 ─────────────────────────────────────────────────────────────
rank_bm25>=0.2.2

# ─── Testing ──────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ─── Monitoring ───────────────────────────────────────────────────────
langsmith>=0.0.63
